html/syntax/parsing-html-fragments/the-input-byte-stream-001.html

<!DOCTYPE html>
<html  lang="en" >
<head>
  <title>HTTP charset</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/html5/syntax.html#the-input-byte-stream'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='http'>
<link rel="stylesheet" type="text/css" href="support/encodingtests-15.css">
</head>
<body>



<div class='test'><div id='box' class='ýäè'>&nbsp;</div></div>


<script>
test(function() {
assert_true(window.getComputedStyle(document.getElementById('box')).getPropertyValue('width') == '100px');
}, "Setting the HTTP header charset declaration will affect the encoding of a page.");
</script>

<div id='log'></div>

</body>
</html>
